Upgrade runtime module
