+++ fastdeploy/backends/sophgo/option.h
@@ -13,6 +13,7 @@
 // limitations under the License.
 #pragma once
 
+#include "fastdeploy/core/fd_type.h"
 #include <cstring>
 #include <iostream>
 #include <memory>
