+++ fastdeploy/fastdeploy_model.h
@@ -121,9 +121,7 @@ class FASTDEPLOY_DECL FastDeployModel {
     std::vector<FDTensor>().swap(reused_output_tensors_);
   }
 
-  virtual fastdeploy::Runtime* CloneRuntime() {
-    return runtime_->Clone();
-  }
+  virtual fastdeploy::Runtime* CloneRuntime() { return runtime_->Clone(); }
 
   virtual bool SetRuntime(fastdeploy::Runtime* clone_runtime) {
     runtime_ = std::unique_ptr<Runtime>(clone_runtime);
